Decouple buffer bindings management from the buffer cache
